+++ src/components/catalog/search/CatalogSearch.scss
@@ -59,15 +59,18 @@
   min-height: 100px;
   display: flex;
   flex-direction: row;
+  height: 150px;
 
   &-container {
     display: flex;
     flex-direction: row;
+    padding-top: 8px !important;
   }
 
   &-image {
     width: 200px;
     height: 150px;
+    flex: none;
   }
 
   &-content {
@@ -82,11 +85,20 @@
       display: flex;
       flex-direction: column;
     }
+
+    &-name-row {
+      display: flex;
+      flex-direction: row;
+
+      .MuiTypography-root {
+        margin-right: 16px;
+      }
+    }
   }
 
   &-name {
     margin: 0;
-    font-size: 2rem;
+    font-size: 1.6rem;
   }
 
   &-chip-container {
@@ -98,6 +110,15 @@
     & .MuiChip-root {
       margin-right: 4px;
       margin-bottom: 4px;
+    }
+  }
+
+  &-features {
+    margin-top: 4px;
+
+    .MuiChip-root {
+      margin-right: 2px;
+      margin-bottom: 2px;
     }
   }
 }
